lorax: Skip rpm database and /boot removal
The standard lorax templates remove the rpm database and most of the
/boot contents. Because of how osbuild runs the scripts we need to keep
the database and /boot around until later.

So this skips removing the rpm database, and skips any command operating
on /boot.

These files can be excluded from the install.img by adding them to the
squashfs/erofs stage's exclude_paths list.

With these changes it is now possible to add a cleanup stage, after the
dracut stage inside the anaconda-tree pipeline, which will remove the
same files that are removed on the Lorax produced boot.iso

eg. add this after the org.osbuild.dracut stage:

        {
          "type": "org.osbuild.lorax-script",
          "options": {
            "path": "99-generic/runtime-cleanup.tmpl",
            "basearch": "x86_64",
            "product": {
              "name": "",
              "version": ""
            },
            "branding": {
              "release": "fedora-release",
              "logos": "fedora-logos"
            }
          }
        },

And add the excluded files/paths to the org.osbuild.squashfs stage:

          "type": "org.osbuild.squashfs",
          "inputs": {
            "tree": {
              "type": "org.osbuild.tree",
              "origin": "org.osbuild.pipeline",
              "references": [
                "name:anaconda-tree"
              ]
            }
          },
          "options": {
            "filename": "images/install.img",
            "exclude_paths": [
              "boot/efi/.*",
              "boot/grub2/.*",
              "boot/config-.*",
              "boot/initramfs-.*",
              "boot/loader/.*",
              "boot/symvers-.*",
              "boot/System.map-.*",
              "usr/lib/sysimage/rpm/.*",
              "var/lib/rpm/.*",
              "var/lib/yum/.*",
              "var/lib/dnf/.*"
            ],
            "compression": {
              "method": "xz",
              "options": {
                "bcj": "x86"
              }
            }
          }
        },