Run upstream unit test during %check
It's beneficial to have the tests running during the build
for early testing with new Python versions which happens continually.

Skip tests on EPEL 8: pytest available there is too old.

Not all test dependencies are availabe in RHEL, filter the toml-writing
library and tests accordingly.

Some tests don't pass on specific architectures, but since osbuild is a
noarch package, we don't have a simple way to filter them out.
The issues are described in the comment, and tests are skipped based on
the target OS.