manifest_tests: allow generating manifests in a container
In order to run manifest tests, we need to download the image build
cache corresponding to the checked out 'osbuild/imags' ref. To determine
the cache to download, we need to generate all test manifests.
Generating test manifests requires having Golang installed of at least
the version required by the checked out 'osbuild/images'. This turns out
to be a problem when we do this on a GA RHEL version with the latest
upstream version of 'osbuild/images'.

This commit mitigates the situation by allowing us to use a container
(i.e. with latest stable Fedora) to generate the testig manifests.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>