docs: add manifest guide
Add a guide on how manifests in osbuild work.

The purpose of this guide is to work as a step by step tutorial of sorts
for how manifests are created.  It guides a reader through hand-writing
simple manifests (even though manifests aren't meant to be written by
hand).  This guide should be useful for anyone working with osbuild
stages or image definitions to understand the building blocks one by
one.