github: avoid single person code owners
When a single person is listed as a code owner for a path, a PR that
modifies files in that path *requires* that person to approve before
merging, which can be a problem when that person is away.  Add
osbuild-reviewers to all paths that previously specified just one owner.

Since Tomáš is already a member of osbuild-reviewers, the effect of
those lines now is that he will always be automatically assigned as a
reviewer of PRs that modify files in those paths, and the rest of the
reviewers are selected randomly from reviewers team.