test: add test that ensures osbuild can be run with plain podman
This test ensures that we can build manifests that do not require
access to loop devices with with rootless podman. We still need
CAP_SYS_ADMIN, CAP_NET_ADMIN inside the container but that is
much less than the previous required rootfull podman.

Note that for non-selinux systems we need to apply a workaround
for the test here we will not be able to build
org.osbuild.selinux with a rootless container because selinux
labeling will not work in a user namespace.