stages/systemd-repart: introduce systemd-repart
This is a stage that allows for creating disk images with
`systemd-repart`. It's set up so it reads any `repart.d` dropin files
from the given root tree and creates a disk image based on that.

In the future we can iterate on this and allow for explicit passing of
dropin files, and implement more arguments.

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>