selinux: Add osbuild_container_t domain for using osbuild in a container
osbuild is sometimes used in a privileged container. To do so you
currently need to run with selinux domain unconfined_t. However, even
then it doesn't work if any file label is set during the image build
that is not part of the host selinux policy. To make that work, the
domain requires the mac_admin capability, which unconfined_t doesn't
allow. The osbuild_t type in the osbuild selinux policy allows this,
but that can't be used in the container.

So, this instead adds a new osbuild_container_t type that can be used
with podman run instead of unconfined_t, allowing osbuild to have the
required capability. However, instead of making osbuild_container_t be
unconfined, we define a tighter set of permissions that seems to make
osbuild work for me.

I do sometimes see one remaining AVC:
```
AVC avc:  denied  { setsched } for  pid=1689360 comm="mount" scontext=system_u:system_r:osbuild_container_t:s0:c440,c895 tcontext=system_u:system_r:kernel_t:s0 tclass=process permissive=0
```

However, this doesn't break the build, and I prefer to not grant
unnecessary permissions.

The new type is in a new policy module (ostree-container), so it can
be installed on the container host with no dependencies on osbuild
packages.