osbuild: handle exception from the _message() in BaseAPI()
This commit fixes the issue that exception from the
`_message()` method (that is provided by the implemenations
of an API like `remoteloop.py`) are not captured. This can
lead to hangs, e.g. if the remoteloop.py sends a message
that a device needs to be created but the creation fails,
then no message is ever sent to the client and the code
hangs. I.e.:
```
    def _message(self, msg, fds, sock):
        ...
        devname = self._create_device(fd, dir_fd, offset, sizelimit, lock, partscan, read_only, sector_size)
        sock.send({"devname": devname})
```
here `self._create_device()` errors so `sock.send()` never
runs. Instead of fixing just this specific instance add a
general `with message_exception_handler()` to the dispatcher
so that any API implementation that raises an exception gets
that communicated back to the client.