schema/metadata: `exports`
Include a new `exports` key which is a list of objects under the
`metadata` object. This allows manifest generators such as `images` to
include information about the pipeline(s) that should be exported.

This would allow for manifests that include this metadata to be rebuilt
with their exports exported automatically, or for `osbuild` to grow a
`--export-defaults` flag to automatically export all pipelines that
match.

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>