osbuild: add/use new `api.get_exception_from_msg()` helper
When a remote command fails in remoteloop.py we currently get
a quite confusing error:
```
  File "/run/osbuild/bin/org.osbuild.rawfs", line 102, in main
    with loop_client.device(image) as loop, mount(loop, mountpoint):
  File "/usr/lib64/python3.12/contextlib.py", line 137, in __enter__
    return next(self.gen)
           ^^^^^^^^^^^^^^
  File "/run/osbuild/lib/osbuild/remoteloop.py", line 138, in device
    path = os.path.join("/dev", payload["devname"])
                                ~~~~~~~^^^^^^^^^^^
KeyError: 'devname'
```
The reason is that there is no "devname" as this is a message of
type "exception". So this code follows the pattern that
objectstore is using to get errors from message now so that we
get the get any remote errors and raise them.