stages: add `efi_src_dir` config to grub2.iso stage
We currently hardcode the location of the efi binaries to
`/boot/efi/EFI/`. This makes sense in general but when working
with bootc containers the location is usually different and
inside `/usr/lib/bootupd/updates/EFI/`. So this commit makes
the location configurable in a  similar way as we already
have it for the grub2 (non-iso) stage (in the grub2 stage
its under `uefi` but in this stage everything is flat
incuding the vendor so `efi_src_dir` is flat as well).

The rational is that we want to build ISOs from bootc base
images (c.f.
https://github.com/osbuild/bootc-image-builder/pull/1053)
but also do not want every bootc installer image to have
a special bootupd grub iso configuration (which would
also hard to maintain as it would be duplicated in every
container). So instead we use this stage to generate the
config and just point to a different EFI dir.