stages: tweak grub2.iso option to support `efi_src_dirs` (plural)
Sadly the support for a single custom `efi_src_dir` [0] to look
for EFI binaries is not enough for the anaconda-bootc work:

The default shim on a bootc system is diverted from
`/boot/efi/EFI/` to `/usr/lib/bootupd/updates/EFI/`.

We also need to install "grub2-iso-x64-cdboot" for a booting
iso from the base bootc container. This will be put under
the regular `/boot/efi/EFI/<vendor>/gcdx64.efi` location.

So the `org.osbuild.grub2.iso` stage will need to be able
to lookup the required binaries in both places. Hence
this commit tweaks the option to allow an array of
paths for the `efi_src_dirs` so that this can be supported.

The images library will then set this to:
```go
bootTreePipeline.EfiSrcDirs = []string{
  "/boot/efi/EFI/",
  "/usr/lib/bootupd/updates/EFI/",
}
```

This is needed for
https://github.com/osbuild/bootc-image-builder/pull/1059

[0] https://github.com/osbuild/osbuild/pull/2202