osbuild: if Loop.mknod() cannot bind mount fallback to mknod
In https://github.com/osbuild/osbuild/pull/2187 we added code
to allow running osbuild with less privileges. However the new
code fails in the test_assemblers.py code when the assemblers
try to request a loop device from inside the bubble wrap container
and Loop.mknod() will not be able to bind mount the device to
the passed dir_fd. The exact reason is unclear, it seems related
to the mount namespace rules [0] but there are some open questions
about this theory.

To fix the regression (and potential future issues) always fallback
to the previous `mknod()` if the bind mount fails.

[0] https://github.com/osbuild/osbuild/pull/2187#issuecomment-3327117870