Allow loopback device to mount files from other pipelines
This change adds support to the org.osbuild.loopback device type for an
'input://FILE' filename that will mount the file from the pipeline
specified in the stage's input section.

It also checks to make sure that if input:// is used on any devices it
MUST be used on all of them and will raise an error if this is not true.
This is because the device mounts are setup before the stage is actually
run and they only have a single parent tree which is either the current
pipeline's tree or the tree from the input pipeline.

This allows mounting disk image partitions for use in later pipelines.
The mounted partitions will be under /run/osbuild/mount while the stage
is running.

eg.

{
  "type": "org.osbuild.example",
  "inputs": {
    "tree": {
      "type": "org.osbuild.tree",
      "origin": "org.osbuild.pipeline",
      "references": [
        "name:image"
      ]
    }
  },
  "options": {
    ... stage options ...
  },
  {
    "devices": {
      "disk": {
        "type": "org.osbuild.loopback",
        "options": {
          "filename": "input://tree/disk.raw",
          "partscan": true,
          "read-only": true
        }
      }
    }
  },
  "mounts": [
    {
      "name": "-",
      "type": "org.osbuild.xfs",
      "source": "disk",
      "target": "/",
      "partition": 4
    }
  ]
}