erofs: Add optional source path
This can be used to allow the erofs stage to be used to compress a
mounted filesystem instead of the input. eg. Input from a previous
pipeline that has a disk image, use devices+mount to mount the
partitions of the image, and source = "mount:///" to compress it.

Supports both 'mount://-/' and 'input://tree/' paths, and maintains
previous behavior if 'source' is not set.