drop remove_signatures from org.osbuild.container-deploy
The original reason to implement this for org.osbuild.container-deploy
was because we were doing `skopeo copy` from one containers-storage
instance to another and we needed the option if the container was
signed.

With the most recent commit to container_mount in containers.py we
don't actually ever skopeo copy from one containers-storage instance
to another, but rather `podman mount` directly, so we should be able
to drop this option.

Context in:
- https://github.com/containers/container-libs/issues/197
- https://github.com/osbuild/bootc-image-builder/issues/681
- https://github.com/osbuild/osbuild/pull/1906