fedora-build-v2: Add VM support to the build pipeline
This adds a minimal initrd + kernel to the build pipeline
that can be used to run a virtiofs-based vm.

The idea is that you run qemu -kernel ../vmlinuz -initrd ../initramfs.img,
and then use the -append feature to add to the kernel commandline:

 * rootfstype=virtiofs root=rootfs - this picks a rootfs from virtio
 * init=/path/to/some/executable

Additionally, the initrd is set up to mount any virtiofs tagged "mnt0"
at /mnt before starting the init. The goal is to pass the osbuild
libdir as this tag, thus allowing init= to be set to a python program
from the ostree sources, while rootfs points to the build pipeline
output.