Add support for serializing stages (and related objects)
This changes the following objects to have to/from_dict() calls that
lets us pass them easily over json.

 * Device
 * Input
 * Mount
 * BuildResult
 * Stage
 * Runner

Also, some of these store `info_path`, which is a path relative to
libdir.  We serialize these relative to libdir so we can deserialize
in a VM where libdir is in a different path.