stages: add new link stage
This introduces a new stage, `org.osbuild.ln`, for creating
links within the filesystem tree.

This takes a list of paths, each with a `target` and a `link_name`
parameters, inspired by the copy stage.

The target is a simple string because we want the link to be either
absolute or relative.