Support running pipeline in a VM
If you specify --in-vm=FOO, then when building the FOO pipeline, we
bring up a qemu instance (using the build pipeline for kernel and
rootfs), and for each stage we serialize it and pass it to the vm
instance via the vm operation "run_stage".

This operation supports logging to a monitor, as well as returning the
serialized BuildResult.

This means all the kernel-related operation in the pipeline such as
loopback mounts, etc, are done in the VM with a kernel that matches
the build image (so the host kernel doesn't influence the build).