tools/test/depsolve: add common check for dump() v1 API
Add a common helper function for checking the structure of the "search"
command response and use it.

We don't need to check every value in the "search" command response for
the purpose of testing. We can check most of the behavior just by
checking the package full NEVRA in the response and that every
package entry contains the expected properties.

Thanks to Michael Vogt for sparking the idea by suggesting to check
package name and the release number.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>