osbuild/solver/dnf: keep some request properties in the Solver object
Assign certain values from the solver request, to the Solver object
properties in its constructor. These are `repos`, `root_dir` and a
calculated `request_repo_ids` properties. This will make it possible to
access them also in `search()` and `dump()` command implementations,
which is needed in order to determine the GPG keys used by repositories
referenced by packages in the command result.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>