Test: move 'repo_servers' fixture to project root conftest.py
I'd like to reuse the 'repo_servers' fixture from the
tools/test/test_depsolve.py, in the Solver unit tests under test/mod/
directory. The only way to do this is to place them into the conftest.py
file in the project root, so that pytest can discover it in all cases.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>