osbuild/testutil/dnf*: accept list of servers in depsolve_pkgset()
Modify the depsolve_pkgset() in testutil for both, DNF4 and DNF5, to
accept list of servers to use, instead of accepting path to the
repository directory and using it as 'file://...'. This will make the
helper function more flexible to use in unit tests.

Adjust all places that use these function to the change.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>