osbuild/testutil/dnf*: set the repo name
Previously, the test functions would configure the provided repository
name as the repo ID. Set it also as the repo name in DNF config. DNF4
defaults to using the ID for the name when not set, but DNF5 doesn't.
Set the repo name explicitly to circumvent this inconsistency in
testing.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>