osbuild/solver: add data classes for Solver API request
Add data classes for Solver API request. This will enable us to also
version and change the Solver request structure, if needed. If nothing
else, the 'api_version' property will have to be added to support
requesting different API version response.

The data classes will enable us to pass typed objects to the Solver
instance methods, instead of a dictionary or an array.

The data classes will enable us to define default values in a
single place, instead of defining it in every Solver implementation.

The data separation into classes is done in logical manner, not
necessarily corresponding to the current request structure, because may
change with different API versions in the future.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>