org.osbuild.kickstart: rewrite mutual exclusion key schema
The next commit is introducing a new bootc verb. Similarly to ostree,
and ostreecontainer, this one is also a part of this mutual exclusive
group. I haven't been be able to find out a way to add a third mutual
exclusive key, so I decided to rewrite the schema a bit to make it
simpler using oneOf. Note that this changes the validation error message.