osbuild/solver/api: add V1 parser for the Solver request
Add V1 API parser for the Solver request. This is the base for
decoupling the Solver response JSON representation from the internal
request data classes. It will make it possible to specify the API
version in the request and also support different versions of the Solver
request at the same time.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>