tools/osbuild-depsolve-dnf: use Solver API request classes and parser
Use the Solver API request data classes and the request parser in the
osbuild-depsolve-dnf. This implies the use of them in all Solver
implementations, which methods now take the classes objects, instead of
dictionaries with the request value. The response serializer no longer
hard-codes the response API version, but uses the same version as the
one specified (parsed) in the Solver API request.

This make the solver able to support multiple API versions at the same
time.

Adjust functional tests to reflect the improved error messages for
invalid API requests.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>