ostree.grub2 update template with boot path
This stage find the default boot path for an ostree tree and substitutes
it for '@OSTREE@' in a grub.cfg file. The config file is specified by
'filename' and the ostree can be either an input or a mount by setting
'source'.

For example, an ostree mounted from a disk image file would look like
this:

{
  "type": "org.osbuild.ostree.grub2",
  "options": {
    "source": "mount://-/",
    "filename": "grub.cfg"
  },
  "devices": {
    "disk": {
      "type": "org.osbuild.loopback",
      "options": {
        "filename": "disk.raw",
        "partscan": true,
        "read-only": true
      }
    }
  },
  "mounts": [
    {
      "name": "-",
      "type": "org.osbuild.xfs",
      "source": "disk",
      "target": "/",
      "partition": 4
    }
  ]
}

Resolves: HMS-9662
Related: HMS-8933