osbuild/solver/api/v1: parse only arguments relevant to the command
Parse only arguments that are relevant to the requested command. I.e. do
not parse the 'search' property in the `arguments` in case the
'depsolve' command is specified. The osbuild/images library sends over
requests with irrelevant fields set to wrong types, which was causing
issues with the new versioned API and stricter validation. This change
fixes the incompatibility.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>