osbuild/solver/request: allow empty pkg specs in depsolve transaction
We can't enforce that the package_specs list in every depsolve
transaction is not empty, because there is a "bug" in the
osbuild/images "os" pipeline implementation. The commit [0]
introduced three transactions in osbuild/images library. However,
the "customizations" package set is added to the list of transactions
unconditionally, even if it is empty. This is the case for "container"
and "wsl" image types. Other image types enable at least SELinux,
which adds the "selinux-policy-targeted" package to the list.

Relax the condition and amend unit tests.

[0] https://github.com/osbuild/images/commit/8b779619aa0c3a9b8537f6bb79324303cb87909c

Signed-off-by: Tomáš Hozza <thozza@redhat.com>