osbuild/solver: include Repositories in the dump and search results
Include the list of Repositories, referenced by the packages in the
command result, within the result. The current API version won't use it,
but future versions should return the list of repositories as well. The
reason is that Packages refer to repo ID that they come from, but the
IDs are usually generated for each Solver request and they are not
human-friendly. Including the repos in the result, and later in the API
response, will make it easy to connect the packages with the
repositories on the caller side.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>