osbuild/solver/api: all serializers expect Solver name as an argument
Include the Solver name in serializer functions for all Solver commands.
Until now, it has been used only for the depsolve command. The current
API version won't use it, but future versions should include this
information in responses for all commands, to make it clear which solver
implementation handled and produced the result.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>