osbuild/solver: always sort depsolve result package list
DNF4 solver returns packages in alphabetical order, while DNF5 solver
returns packages in topological order. The order is not important for
our use case, it just needs to be deterministic and stable.
The org.osbuild.rpm stage as generated by osbuild/images does not
depend on the order of packages, because rpm gets the full package
set at once and it will reorder the packages as needed when
installing.

Since we use only DNF4 solver by default in Fedora and RHEL, we
currently get alphabetically ordered package sets from the Solver by
default. So this is the de facto standard order.

Explicitly sort returned package set alphabetically by full NEVRA in
both Solver implementations. In DNF4 to not implicitly rely on the DNF
implementation details. In DNF5 to be consistent with DNF4.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>