osbuild/solver: make Solver completely unaware of any external API
Move all API serialization and API version handling from the Solver
implementation to tools/osbuild-depsolve-dnf. The need to construct
Solver with a request.SolverRequest object, which must have API version
specified, turned out to be a weird requirement while writing unit tests
for the Solver. This separation makes also logical sense, since the
Solver does not need to know anything about any external API. It accepts
data classes and produces results using data classes, where all of them
are completely API version agnostic.

As a side effect, the DNF5 Solver no longer collects excluded packages
list from the depsolve command arguments in its constructor. The reason
is that no command arguments are passed any more to the Solver
constructor. The implementation was already arguably broken in this
regard in multiple ways, so doing so does not make things worse.
This whole part needs to be reworked in the future anyway.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>