osbuild/solver/dnf5: set repo name to repo ID if no name is provided
DNF4 defaults to using the repo ID as the repo name, in case no explicit
repo name is provided when configuring repositories. When we later
gather repo configs from a dumped/depsolved/searched package set, the
repo name is always set.

DNF5 on the contrary keeps the repo name empty, unless it is explicitly
set in the repo config. Based on my discussion with the DNF team, this
is most probably a bug [1].

Workaround the inconsistency by explicitly setting repo name to the repo
ID.

[1] https://github.com/rpm-software-management/dnf5/issues/2533

Signed-off-by: Tomáš Hozza <thozza@redhat.com>