in-vm: Mount osbuild python path on /mnt and libdir separately
It turns out the --in-vm code didn't work when packaged in an rpm
because then the osbuild libdir and the python code as not in the
same place, so when mounting libdir at /mnt it could not find vm.py for
the init process.

We now always mount the python base path where the osbuild module
lives at /mnt, and then we mount the libdir separately. This works
both when running as an rpm, and when running from the source tree.