osbuild/solver/model: add 'rhsm' option to Repository
Add the 'rhsm' option to the model.Repository. The idea is that the
Solver will do the RHSM secrets discovery and set the appropriate
Repository properties to the host system secrets. Similarly, this
'rhsm' value should be set to True if the same repo ID used the RHSM
secrets from the host. The value should be then sent back to the
depsolver API caller to make use of it when generating manifest.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>