ostree.grub2: ostree requires leading /
ostree requires the leading /, if it is missing it will still boot but
its internal regex to match the ostree target will fail and it will not
generate the var.mount unit needed to setup /var. A side effect of this
is a missing /var/home/user directory.

Related: HMS-9662
Related: HMS-8933