osbuild/util/rhsm: make Subscriptions.get_secrets() take list of URLs
Given the way get_secrets() is implemented, it is hard to call it
multiple times for each baseurl or subsequently for metalink or
mirrorlist. Specifically, because if the passed URL does not match any
redhat.repo URLs, the function falls back to default secrets or raises
a RuntimeError if these are not found. Therefore it is hard to tell for
the caller if there was an URL match, or if the returned secrets are the
fall-back secrets.

Modify the get_secrets() method to accept list of URLs. It iterates over
them and returns secrets for the first matching URL. In case of no
match, the old behavior is preserved.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>