osbuild/solver: make SolverBase handle 'rhsm' flag in repo config
Extend the SolverBase to handle the RHSM secrets discovery for
RepositoryConfig instances from the request. If a repository has the
rhsm=True, the Solver will assign the sslcacert, sslclientkey and
sslclientcert to the discovered host RHSM secrets.

Related to the secrets discovery, add a helper
function set_rhsm_flag(). It should be used by a Solver
implementation to mark Repository objects created from a DNF API
response as rhsm=True, in case the Solver assigned the host system
RHSM secrets to a repo with the same ID. This may be then used by
an API implementation to decide if the host RHSM secrets should be
omitted from the API response.

Add comprehensive unit tests covering the new functionality.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>