osbuild/solver/dnf*: set all model.Repository fields
In Solver implementations, set all fields that model.Repository
contains, when converting the DNF-specific representation of repo
configuration to the model representation.

It makes sense to get the same model.Repository from a Solver command
implementation, that was passed to the Solver when constructed.

For API simplicity and consistency, convert the metadata_expire value
to a string with the second unit specifier.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>