tools/test/depsolve: add V2 package and repository validators
Add assert_package_v2() and assert_repository_v2() helper functions
to validate the structure of V2 API response objects. The nice thing
about V2 API is that the Package and Repository objects are represented
in exactly the same way in responses from all commands. This makes it
possible to have a common helper functions to validate them.

These validators will be used by the command-specific V2 response
validators in the next step.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>