runners: set up sequoia on RHEL
With RHEL 10.1 z-stream a new cryptographic key is used to sign RPMs.
This means that RPM versions that understand multiple signatures will
try to verify through sequoia that this key is valid.

When verifying packages post-buildroot (so using an RPM version from the
buildroot which understands these multi-signatures) sequoia-gpg will be
used to validate. We don't set up sequoia-gpg and it needs its config to
understand the new signature as its not in the defaults.

Change this so we always set up crypto-policies so sequoia knows what to
use instead of its defaults.

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>