GHA: add workflow to update osbuild-ci container image references
Add a GHA, which will scan all GHAs in the repository for use of
'osbuild-ci*' container images and update the references to the latest
tag available. The latest tag can be passed as an argument or it will be
automatically determined by the script. The script also ensures that the
image reference that it plans to use points to an existing image, to
prevent breaking the workflow. The report of changes made can be
optionally written into a file for an easy PR description inclusion.

The idea is that this workflow will be possible to trigger manually or
externally from a different repo. The second scenario is particularly
useful, since it will allow the osbuild/containers repo to notify this
repository when new images are built and deployed to the container
registry.

This will ensure that we keep all of our 'osbuild-ci*' images in sync
and updated.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>