stages/rpm: add option to ignore key import errors
This is needed because RHEL 9's rpm does not understand PQC keys, but
the RHEL 9 packages are still signed with PQC. This means that when
building the RHEL 9 buildroot on RHEL 10 the PQC keys need to be
imported successfully, but when building the RHEL 9 OS pipeline inside
of the RHEL 9 buildroot, the import errors need to be ignored.

`rpmkeys` will fail to import keys it does not understand
(e.g. post-quantum keys). Note that rpmkeys does not stop when it
encounters a key it does not understand. So just setting `check=` to
False is enough. [1]

We assume that if rpm actually requires them for checking the
signatures, it will fail during `rpmkeys --checksig`.

[1]: https://github.com/rpm-software-management/rpm/blob/4a9b7b5908d8b463a836b51322242677677bd8b7/lib/rpmchecksig.cc#L66
(there is no break statement in the error branch).